refactor: reuse launch lifecycle helpers
